Fixed state persistance to localStorage.
State is now correctly being stored and hydrated from localStorage.
However, the state machine does not currently resume on the step
you were on when the tab was closed.  Need to dig more into the xstate
docs to figure out how to do that without ending up where we were
before.