Fixed scoring for single character genes.
This commit fixes scoring when searches are for
genes with a single character symbol.  It does so
by utilizing a trigram index to power like queries
and to calculate a distance between the query and
the symbol.  Searches across symbols and all other
ID/symbol/synonym fields are then sorted by this
distance (ascending) and scores of exact and
wildcard matches against the other fields.